Refactor: 활동점수 상세 모달 로직 modal/* 아래로 이동
