+++ src/components/ActivityScore/index.ts
@@ -1,6 +1,5 @@
 export { default as PersonalInfoCard } from './PersonalInfoCard/PersonalInfoCard.component';
 export { default as ScoreCard } from './ScoreCard/ScoreCard.component';
 export { default as Icon } from './Icon/Icon.component';
-export { default as ActivityScoreModalDialog } from './ActivityScoreModalDialog/ActivityScoreModalDialog.component';
 export * from './utils';
 export * from './constants';
