+++ src/components/modal/index.ts
@@ -2,3 +2,4 @@ export { default as ChangeResultModalDialog } from './ChangeResultModalDialog/Ch
 export { default as SmsSendDetailInfoModalDialog } from './SmsSendDetailInfoModalDialog/SmsSendDetailInfoModalDialog.component';
 export { default as SmsSendDetailListModalDialog } from './SmsSendDetailListModalDialog/SmsSendDetailListModalDialog.component';
 export { default as ApplyActivityScoreModalDialog } from './ApplyActivityScoreModalDialog/ApplyActivityScoreModalDialog.component';
+export { default as ActivityScoreModalDialog } from './ActivityScoreModalDialog/ActivityScoreModalDialog.component';
