+++ src/pages/ActivityScoreDetail/ActivityScoreDetail.page.tsx
@@ -3,10 +3,9 @@ import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { useParams } from 'react-router-dom';
 import { BackButton, Button, Table } from '@/components';
 import * as Styled from './ActivityScoreDetail.styled';
-import { useHistory, useToggleState } from '@/hooks';
+import { useHistory } from '@/hooks';
 import { PATH } from '@/constants';
 import {
-  ActivityScoreModalDialog,
   Icon,
   PersonalInfoCard,
   RangeType,
@@ -36,11 +35,14 @@ const getScoreRangeType = (score: number) => {
 
 const ActivityScoreDetail = () => {
   const { handleGoBack } = useHistory();
-  const [isActivityScoreModalOpened, toggleActivityScoreModalOpened] = useToggleState(false);
   const { generationNumber: generationNumberParam, memberId: memberIdParam } = useParams();
 
   const handleApplyActivityScoreModal = useSetRecoilState(
     $modalByStorage(ModalKey.applyActivityScoreModalDialog),
+  );
+
+  const handleActivityScoreModal = useSetRecoilState(
+    $modalByStorage(ModalKey.activityScoreModalDialog),
   );
 
   const columns: TableColumn<ScoreHistory>[] = [
@@ -63,7 +65,16 @@ const ActivityScoreDetail = () => {
 
         return (
           <>
-            <Styled.ActivityTitle isCanceled={isCanceled} onClick={toggleActivityScoreModalOpened}>
+            <Styled.ActivityTitle
+              isCanceled={isCanceled}
+              onClick={() =>
+                handleActivityScoreModal({
+                  key: ModalKey.activityScoreModalDialog,
+                  isOpen: true,
+                  props: {},
+                })
+              }
+            >
               {scoreName}
             </Styled.ActivityTitle>
             {isCanceled && <Styled.CancelLabel>취소</Styled.CancelLabel>}
@@ -125,50 +136,45 @@ const ActivityScoreDetail = () => {
     );
 
   return (
-    <>
-      <Styled.ActivityScoreDetailPage>
-        <BackButton label="목록 돌아가기" onClick={() => handleGoBack(PATH.ACTIVITY_SCORE)} />
-        <Styled.Headline>활동점수 상세</Styled.Headline>
-        <Styled.Row>
-          <PersonalInfoCard
-            name={name}
-            platform={platform}
-            identification={identification}
-            generationNumber={generationNumber}
+    <Styled.ActivityScoreDetailPage>
+      <BackButton label="목록 돌아가기" onClick={() => handleGoBack(PATH.ACTIVITY_SCORE)} />
+      <Styled.Headline>활동점수 상세</Styled.Headline>
+      <Styled.Row>
+        <PersonalInfoCard
+          name={name}
+          platform={platform}
+          identification={identification}
+          generationNumber={generationNumber}
+        />
+        <ScoreCard totalScore={totalScore} />
+      </Styled.Row>
+      <Styled.Content>
+        <Styled.ContentHeader>
+          <h3>활동점수 히스토리</h3>
+          <Button
+            shape={ButtonShape.defaultLine}
+            label="점수 추가"
+            Icon={Plus}
+            onClick={() => {
+              handleApplyActivityScoreModal({
+                key: ModalKey.applyActivityScoreModalDialog,
+                isOpen: true,
+                props: {
+                  generationNumber: parseUrlParam(generationNumberParam),
+                  memberId: parseUrlParam(memberIdParam),
+                },
+              });
+            }}
           />
-          <ScoreCard totalScore={totalScore} />
-        </Styled.Row>
-        <Styled.Content>
-          <Styled.ContentHeader>
-            <h3>활동점수 히스토리</h3>
-            <Button
-              shape={ButtonShape.defaultLine}
-              label="점수 추가"
-              Icon={Plus}
-              onClick={() => {
-                handleApplyActivityScoreModal({
-                  key: ModalKey.applyActivityScoreModalDialog,
-                  isOpen: true,
-                  props: {
-                    generationNumber: parseUrlParam(generationNumberParam),
-                    memberId: parseUrlParam(memberIdParam),
-                  },
-                });
-              }}
-            />
-          </Styled.ContentHeader>
-          <Table
-            prefix="score-history"
-            columns={columns}
-            rows={scoreHistoryResponses}
-            supportBar={{}}
-          />
-        </Styled.Content>
-      </Styled.ActivityScoreDetailPage>
-      {isActivityScoreModalOpened && (
-        <ActivityScoreModalDialog onClose={toggleActivityScoreModalOpened} />
-      )}
-    </>
+        </Styled.ContentHeader>
+        <Table
+          prefix="score-history"
+          columns={columns}
+          rows={scoreHistoryResponses}
+          supportBar={{}}
+        />
+      </Styled.Content>
+    </Styled.ActivityScoreDetailPage>
   );
 };
