+++ src/store/modal.ts
@@ -8,6 +8,7 @@ import { SmsSendModalDialogProps } from '@/components/common/SmsSendModalDialog/
 import { SmsSendDetailInfoModalDialogProps } from '@/components/modal/SmsSendDetailInfoModalDialog/SmsSendDetailInfoModalDialog.component';
 import { SmsSendDetailListModalDialogProps } from '@/components/modal/SmsSendDetailListModalDialog/SmsSendDetailListModalDialog.component';
 import { ApplyActivityScoreModalDialogProps } from '@/components/modal/ApplyActivityScoreModalDialog/ApplyActivityScoreModalDialog.component';
+import { ActivityScoreModalDialogProps } from '@/components/modal/ActivityScoreModalDialog/ActivityScoreModalDialog.component';
 
 export const ModalKey = {
   alertModalDialog: 'alertModalDialog',
@@ -16,6 +17,7 @@ export const ModalKey = {
   smsSendDetailInfoModalDialog: 'smsSendDetailInfoModalDialog',
   smsSendDetailListModalDialog: 'smsSendDetailListModalDialog',
   applyActivityScoreModalDialog: 'applyActivityScoreModalDialog',
+  activityScoreModalDialog: 'activityScoreModalDialog',
 } as const;
 
 export type ModalKeyType = ValueOf<typeof ModalKey>;
@@ -25,7 +27,8 @@ export type ModalProps =
   | SmsSendDetailListModalDialogProps
   | SmsSendDetailInfoModalDialogProps
   | ChangeResultModalDialogProps
-  | ApplyActivityScoreModalDialogProps;
+  | ApplyActivityScoreModalDialogProps
+  | ActivityScoreModalDialogProps;
 
 export interface Modal {
   key: ModalKeyType;
